also expose server_response_reader (to be used by CLI)
